install antd and replace modal
