+++ package.json
@@ -14,6 +14,7 @@
     "@fortawesome/free-regular-svg-icons": "^6.3.0",
     "@fortawesome/free-solid-svg-icons": "^6.3.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "antd": "^5.4.0",
     "babel-plugin-macros": "^3.1.0",
     "fast-average-color": "^9.3.0",
     "framer-motion": "^10.8.5",
