+++ src/index.css
@@ -2,7 +2,7 @@
 @tailwind components;
 @tailwind utilities;
 
-@layer base {
+/* @layer base {
   input, select {
     @apply dark:bg-zinc-700 bg-white
   }
@@ -15,7 +15,7 @@
   a {
     @apply outline-none;
   }
-}
+} */
 
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
